Arudino Code v1.0
In this study, using a wireless network, remote control of a mobile robot with a camera mounted on it was performed automatically, using the keyboard arrow keys, and on the way drawn on a virtual map. The purpose of designing this mobile robot is to create an open platform for developing applications that can perform more complex tasks. The mobile robot platform has a wireless communication interface that allows control over the internet using C # compatible devices. This feature ensures that the robot has a flexible structure.